requirements.txt
decorator==4.0.4
Django==1.8.4
django-cors-headers==1.1.0
django-hstore==1.4
djangorestframework==3.2.4
djangorestframework-hstore==1.3
ipdb==0.8.1
ipython==4.0.0
ipython-genutils==0.1.0
Markdown==2.6.2
path.py==8.1.1
pexpect==3.3
pickleshare==0.5
Pillow==3.0.0
psycopg2==2.6.1
python-dateutil==2.4.2
redis==2.10.3
simplegeneric==0.8.1
six==1.9.0
sockjs-tornado==1.0.2
SwampDragon==0.4.2.2
tornado==4.2.1
tornado-redis==2.4.18
traitlets==4.0.0
wheel==0.24.0
